Organize files, add module resolver
