+++ .babelrc
@@ -1,6 +1,7 @@
 {
   "presets": ["stage-2", "es2015"],
   "plugins": [
-    "add-module-exports"
+    "add-module-exports",
+    ["module-resolver", { "root": ["./src"] }]
   ]
 }
